backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.11.7
pydantic-settings==2.1.0
supabase==2.16.0
qdrant-client==1.7.0
requests==2.31.0
PyPDF2==3.0.1
python-docx==1.1.0
langchain-text-splitters==0.0.1
python-jose[cryptography]==3.3.0
python-dotenv==1.0.0
httpx==0.24.1
tenacity==8.2.3
google-generativeai==0.3.2
APScheduler==3.10.4
aiofiles==23.2.1
